Move the OAuthWorker Class to a separate file
